requirements.txt
fastapi
requests
bs4
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
alembic
telebot